question-base/requirements.txt
# Requirements for Industry 4.0 Question Base Tools
# Install with: pip install -r requirements.txt

# Core dependencies
python-docx>=0.8.11        # Read and write Microsoft Word .docx files
jsonschema>=4.17.0         # JSON Schema validation
openpyxl>=3.0.10          # Read/write Excel files (for future features)

# Optional dependencies for enhanced features
pyyaml>=6.0               # YAML support (alternative format)
pandas>=1.5.0             # Data analysis (for statistics and reporting)
tabulate>=0.9.0           # Pretty-print tabular data
colorama>=0.4.6           # Colored terminal output (cross-platform)

# Development dependencies (optional)
pytest>=7.2.0             # Testing framework
black>=22.10.0            # Code formatter
pylint>=2.15.0            # Linting
mypy>=0.991               # Type checking
